feat(skipWhile): add higher-order lettable version of skipWhile
